requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
SQLAlchemy>=1.4
python-multipart
aiohttp
pydantic
